# swagger.yml
openapi: 3.0.0
info:
  title: Ecommerce Backend CH
  description: Simulador de e-commerce, proyecto para el curso Backend de CoderHouse.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
  - url: http://acaIriaHeroku
    description: Optional server description, e.g. Internal staging server for testing
...
paths:
  /api/users/login:
    post:
      summary: Login
      tags:
        - Users
      requestBody:
        required: true
        content:
          aplication/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
            example:
              email: esteban@gmail.com
              password: asdasd
      description: Verifica la información y permite loguearse
      responses:
        "200":
          description: Logueo autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string

        "401":
          description: Login no autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
  /api/users/signup:
    post:
      summary: Signup
      tags:
        - Users
      requestBody:
        required: true
        content:
          aplication/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                phone:
                  type: number
                adress:
                  type: string
                admin:
                  type: boolean
            example:
              email: esteban@gmail.com
              password: asdasd
              name: Esteban
              phone: 54955412561
              adress: home 1212
              admin: false
      description: Permite registrarse en el sitio
      responses:
        "200":
          description: Logueo autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string

        "400":
          description: Email ya usado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string

        "401":
          description: Body no valido
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
...